Fix Daedalus build script so it copies everything required PSPModel now returns the model.. Fix some compiler warnings.. Small speedup

+++ build_daedalus.sh
@@ -2,30 +2,48 @@
 
 
 function usage() {
-echo "Usage ./build_daedalus.sh BUILD_TYPE"
-echo "Build Types:"
-echo "PSP Release = PSP_RELEASE"
-echo "PSP Debug = PSP_DEBUG"
-echo "Linux Release = LINUX_RELEASE"
-echo "Mac Release = MAC_RELEASE"
-exit
+    echo "Usage ./build_daedalus.sh BUILD_TYPE"
+    echo "Build Types:"
+    echo "PSP Release = PSP_RELEASE"
+    echo "PSP Debug = PSP_DEBUG"
+    echo "Linux Release = LINUX_RELEASE"
+    echo "Mac Release = MAC_RELEASE"
+    exit
 }
 
-function cleanup(){
-  rm -r "$PWD/daedbuild"
-  mkdir "$PWD/daedbuild"
+function pre_prep(){
+
+    if [ -d $PWD/daedbuild ]; then
+        echo "Removing previous build attempt"
+        rm -r "$PWD/daedbuild"
+        mkdir "$PWD/daedbuild"
+    fi 
+  
+    if [ -d $PWD/DaedalusX64 ]; then
+        rm -r $PWD/DaedalusX64/EBOOT.PBP
+    else
+        mkdir $PWD/DaedalusX64
+        mkdir ../DaedalusX64/SaveStates
+        mkdir ../DaedalusX64/SaveGames
+        mkdir ../DaedalusX64/Roms
+    fi
+
 }
 
 function finalPrep() {
-  mkdir ../DaedalusX64
-  mkdir ../DaedalusX64/SaveStates
-  mkdir ../DaedalusX64/SaveGames
-  mkdir ../DaedalusX64/Roms
-    if [ "$1" = PSP_RELEASE ] || [ "$1" = PSP_DEBUG ]; then
-      cp -R ../Data/PSP/* ../DaedalusX64
+
+    if [ ! -d ../DaedalusX64 ]; then
+        mkdir ../DaedalusX64/SaveStates
+        mkdir ../DaedalusX64/SaveGames
+        mkdir ../DaedalusX64/Roms
+    fi
+
+    if [ -f "$PWD/EBOOT.PBP" ]; then
+        mv "$PWD/EBOOT.PBP" ../DaedalusX64/
+        cp -r ../Data/PSP/* ../DaedalusX64/
     else
-      cp -R ../Data/PC/* ../DaedalusX64
-      cp ../Source/SysGL/HLEGraphics/n64.psh ../DaedalusX64
+        cp -r ../Data/PC/* ../DaedalusX64/
+        cp ../Source/SysGL/HLEGraphics/n64.psh ../DaedalusX64
     fi
 }
 
@@ -37,7 +55,6 @@ function buildPSP() {
   make -C "$PWD/../Source/SysPSP/MediaEnginePRX"
 
   make
-  finalPrep
   #No point continuing if the elf file doesn't exist
   if [ -f "$PWD/daedalus.elf" ]; then
     #Pack PBP
@@ -46,7 +63,7 @@ function buildPSP() {
   psp-prxgen daedalus.elf daedalus.prx
   cp ../Source/SysPSP/Resources/eboot_icons/* "$PWD"
   pack-pbp EBOOT.PBP PARAM.SFO icon0.png NULL NULL pic1.png NULL daedalus.prx NULL
-  mv EBOOT.PBP ../DaedalusX64
+  finalPrep
 
 
 fi
@@ -56,13 +73,13 @@ fi
 
 
 if [ "$1" = "PSP_RELEASE" ] || [ "$1" = "PSP_DEBUG" ]; then
-  cleanup
+  pre_prep
     cd "$PWD/daedbuild"
 cmake -DCMAKE_TOOLCHAIN_FILE=../Tools/psptoolchain.cmake -D"$1=1" ../Source
 buildPSP
 
 elif [ "$1" = "LINUX_RELEASE" ] || [ "$1" = "MAC_RELEASE" ]; then
-  cleanup
+  pre_prep
   cd "$PWD/daedbuild"
   cmake -D"$1=1" ../Source
 make
